Drafted first release candidate for 1.4.0

+++ src/engine/src/engine/ViewParser.cpp
@@ -1,80 +1,80 @@
 // Copyright 2022 The SQLNet Company GmbH
-// 
-// This file is licensed under the Elastic License 2.0 (ELv2). 
-// Refer to the LICENSE.txt file in the root of the repository 
+//
+// This file is licensed under the Elastic License 2.0 (ELv2).
+// Refer to the LICENSE.txt file in the root of the repository
 // for details.
-// 
+//
 
 #include "engine/handlers/ViewParser.hpp"
 
-// ----------------------------------------------------------------------------
-
+#include "commands/Fingerprint.hpp"
+#include "containers/DataFrameContent.hpp"
 #include "engine/handlers/BoolOpParser.hpp"
-#include "engine/handlers/CatOpParser.hpp"
-#include "engine/handlers/NumOpParser.hpp"
-
-// ----------------------------------------------------------------------------
+#include "engine/handlers/FloatOpParser.hpp"
+#include "engine/handlers/StringOpParser.hpp"
+#include "engine/utils/Getter.hpp"
+#include "helpers/Aggregations.hpp"
 
 namespace engine {
 namespace handlers {
 
-void ViewParser::add_column(const Poco::JSON::Object& _obj,
-                            containers::DataFrame* _df) {
-  if (!_obj.has("added_")) {
+void ViewParser::add_column(const ViewOp& _cmd,
+                            containers::DataFrame* _df) const {
+  if (!_cmd.get<"added_">()) {
     return;
   }
 
-  const auto added = *JSON::get_object(_obj, "added_");
+  const auto handle = [this, &_cmd, _df](const auto& _json_col) {
+    using Type = std::decay_t<decltype(_json_col)>;
 
-  const auto json_col = *JSON::get_object(added, "col_");
+    const auto& added = *_cmd.get<"added_">();
 
-  const auto name = JSON::get_value<std::string>(added, "name_");
+    const auto& name = added.get<"name_">();
 
-  const auto role = JSON::get_value<std::string>(added, "role_");
+    const auto& role = added.get<"role_">();
 
-  const auto subroles =
-      JSON::array_to_vector<std::string>(JSON::get_array(added, "subroles_"));
+    const auto& subroles = added.get<"subroles_">();
 
-  const auto unit = JSON::get_value<std::string>(added, "unit_");
+    const auto& unit = added.get<"unit_">();
 
-  const auto type = JSON::get_value<std::string>(json_col, "type_");
+    if constexpr (std::is_same<Type,
+                               commands::FloatColumnOrFloatColumnView>()) {
+      const auto column_view =
+          FloatOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-  if (type == FLOAT_COLUMN || type == FLOAT_COLUMN_VIEW) {
-    const auto column_view =
-        NumOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      auto col = column_view.to_column(0, _df->nrows(), true);
 
-    auto col = column_view.to_column(0, _df->nrows(), true);
+      col.set_name(name);
 
-    col.set_name(name);
+      col.set_subroles(subroles);
 
-    col.set_subroles(subroles);
+      col.set_unit(unit);
 
-    col.set_unit(unit);
+      _df->add_float_column(col, role);
+    } else {
+      const auto column_view =
+          StringOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-    _df->add_float_column(col, role);
-  }
+      const auto vec = column_view.to_vector(0, _df->nrows(), true);
 
-  if (type == STRING_COLUMN || type == STRING_COLUMN_VIEW) {
-    const auto column_view =
-        CatOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      assert_true(vec);
 
-    const auto vec = column_view.to_vector(0, _df->nrows(), true);
+      if (role == containers::DataFrame::ROLE_CATEGORICAL ||
+          role == containers::DataFrame::ROLE_JOIN_KEY) {
+        add_int_column_to_df(name, role, subroles, unit, *vec, _df);
+      }
 
-    assert_true(vec);
-
-    if (role == containers::DataFrame::ROLE_CATEGORICAL ||
-        role == containers::DataFrame::ROLE_JOIN_KEY) {
-      add_int_column_to_df(name, role, subroles, unit, *vec, _df);
+      if (role == containers::DataFrame::ROLE_UNUSED ||
+          role == containers::DataFrame::ROLE_TEXT ||
+          role == containers::DataFrame::ROLE_UNUSED_STRING) {
+        add_string_column_to_df(name, role, subroles, unit, *vec, _df);
+      }
     }
+  };
 
-    if (role == containers::DataFrame::ROLE_UNUSED ||
-        role == containers::DataFrame::ROLE_TEXT ||
-        role == containers::DataFrame::ROLE_UNUSED_STRING) {
-      add_string_column_to_df(name, role, subroles, unit, *vec, _df);
-    }
-  }
+  std::visit(handle, _cmd.get<"added_">()->get<"col_">());
 }
 
 // ------------------------------------------------------------------------
@@ -84,7 +84,7 @@ void ViewParser::add_int_column_to_df(const std::string& _name,
                                       const std::vector<std::string>& _subroles,
                                       const std::string& _unit,
                                       const std::vector<strings::String>& _vec,
-                                      containers::DataFrame* _df) {
+                                      containers::DataFrame* _df) const {
   const auto encoding = _role == containers::DataFrame::ROLE_CATEGORICAL
                             ? categories_
                             : join_keys_encoding_;
@@ -113,7 +113,8 @@ void ViewParser::add_int_column_to_df(const std::string& _name,
 void ViewParser::add_string_column_to_df(
     const std::string& _name, const std::string& _role,
     const std::vector<std::string>& _subroles, const std::string& _unit,
-    const std::vector<strings::String>& _vec, containers::DataFrame* _df) {
+    const std::vector<strings::String>& _vec,
+    containers::DataFrame* _df) const {
   auto col = containers::Column<strings::String>(_df->pool());
 
   for (size_t i = 0; i < _vec.size(); ++i) {
@@ -131,10 +132,13 @@ void ViewParser::add_string_column_to_df(
 
 // ----------------------------------------------------------------------------
 
-void ViewParser::drop_columns(const Poco::JSON::Object& _obj,
+void ViewParser::drop_columns(const ViewOp& _cmd,
                               containers::DataFrame* _df) const {
-  const auto dropped =
-      JSON::array_to_vector<std::string>(JSON::get_array(_obj, "dropped_"));
+  if (!_cmd.get<"dropped_">()) {
+    return;
+  }
+
+  const auto& dropped = *_cmd.get<"dropped_">();
 
   for (const auto& name : dropped) {
     _df->remove_column(name);
@@ -144,48 +148,28 @@ void ViewParser::drop_columns(const Poco::JSON::Object& _obj,
 // ----------------------------------------------------------------------------
 
 typename ViewParser::ColumnViewVariant ViewParser::make_column_view(
-    Poco::JSON::Object::Ptr _ptr) const {
-  assert_true(_ptr);
+    const ViewCol& _view_col) const {
+  const auto handle = [this](const auto& _col) -> ColumnViewVariant {
+    using Type = std::decay_t<decltype(_col)>;
 
-  const auto type = JSON::get_value<std::string>(*_ptr, "type_");
-
-  if (type == STRING_COLUMN || type == STRING_COLUMN_VIEW) {
-    return CatOpParser(categories_, join_keys_encoding_, data_frames_)
-        .parse(*_ptr);
-  }
-
-  return NumOpParser(categories_, join_keys_encoding_, data_frames_)
-      .parse(*_ptr);
-}
-
-// ----------------------------------------------------------------------------
-
-Poco::JSON::Array::Ptr ViewParser::make_data(
-    const std::vector<std::vector<std::string>>& _string_vectors) const {
-  auto data = Poco::JSON::Array::Ptr(new Poco::JSON::Array());
-
-  if (_string_vectors.size() == 0) {
-    return data;
-  }
-
-  for (size_t i = 0; i < _string_vectors.at(0).size(); ++i) {
-    auto row = Poco::JSON::Array::Ptr(new Poco::JSON::Array());
-
-    for (size_t j = 0; j < _string_vectors.size(); ++j) {
-      if (i == 0 &&
-          _string_vectors.at(j).size() != _string_vectors.at(0).size()) {
-        throw std::runtime_error(
-            "ViewParser: Resulting columns do not all have "
-            "the same length!");
-      }
-
-      row->add(_string_vectors.at(j).at(i));
+    if constexpr (std::is_same<
+                      Type, typename commands::StringColumnOrStringColumnView::
+                                NamedTupleType>()) {
+      const auto col = commands::StringColumnOrStringColumnView(_col);
+      return StringOpParser(categories_, join_keys_encoding_, data_frames_)
+          .parse(col);
     }
 
-    data->add(row);
-  }
+    if constexpr (std::is_same<Type,
+                               typename commands::FloatColumnOrFloatColumnView::
+                                   NamedTupleType>()) {
+      const auto col = commands::FloatColumnOrFloatColumnView(_col);
+      return FloatOpParser(categories_, join_keys_encoding_, data_frames_)
+          .parse(col);
+    }
+  };
 
-  return data;
+  return std::visit(handle, _view_col);
 }
 
 // ----------------------------------------------------------------------------
@@ -270,8 +254,8 @@ std::vector<std::string> ViewParser::make_string_vector(
       return _str.str();
     };
 
-    return fct::collect::vector<std::string>(
-        *str_col.to_vector(_start, _length, false) | VIEWS::transform(to_str));
+    return fct::collect::vector(*str_col.to_vector(_start, _length, false) |
+                                VIEWS::transform(to_str));
   }
 
   const auto float_col = std::get<containers::ColumnView<Float>>(_column_view);
@@ -279,30 +263,24 @@ std::vector<std::string> ViewParser::make_string_vector(
   const auto float_vec = *float_col.to_vector(_start, _length, false);
 
   if (float_col.unit().find("time stamp") != std::string::npos) {
-    return fct::collect::vector<std::string>(
-        float_vec | VIEWS::transform(io::Parser::ts_to_string));
+    return fct::collect::vector(float_vec |
+                                VIEWS::transform(io::Parser::ts_to_string));
   }
 
   const auto to_string = [](const Float _val) {
     return io::Parser::to_string(_val);
   };
 
-  return fct::collect::vector<std::string>(float_vec |
-                                           VIEWS::transform(to_string));
+  return fct::collect::vector(float_vec | VIEWS::transform(to_string));
 }
 
 // ----------------------------------------------------------------------------
 
-Poco::JSON::Object ViewParser::get_content(
+containers::ViewContent ViewParser::get_content(
     const size_t _draw, const size_t _start, const size_t _length,
-    const bool _force_nrows, const Poco::JSON::Array::Ptr& _cols) const {
-  const auto get_object = [_cols](const size_t _i) {
-    return _cols->getObject(_i);
-  };
-
-  const auto to_column_view =
-      [this](Poco::JSON::Object::Ptr _ptr) -> ColumnViewVariant {
-    return make_column_view(_ptr);
+    const bool _force_nrows, const std::vector<ViewCol>& _cols) const {
+  const auto to_column_view = [this](const auto& _col) -> ColumnViewVariant {
+    return make_column_view(_col);
   };
 
   const auto to_string_vector =
@@ -311,88 +289,95 @@ Poco::JSON::Object ViewParser::get_content(
     return make_string_vector(_start, _length, _column_view);
   };
 
-  const auto make_column_view = fct::compose(get_object, to_column_view);
+  const auto transpose = [](const auto& _data) {
+    if (_data.size() == 0) {
+      return _data;
+    }
+    std::vector<std::vector<std::string>> result;
+    for (size_t i = 0; i < _data.at(0).size(); ++i) {
+      std::vector<std::string> row;
+      for (const auto& _col : _data) {
+        assert_true(_col.size() == _data.at(0).size());
+        row.push_back(_col.at(i));
+      }
+      result.push_back(row);
+    }
+    return result;
+  };
 
-  const auto iota = fct::iota<size_t>(0, _cols->size());
+  const auto column_views =
+      fct::collect::vector(_cols | VIEWS::transform(to_column_view));
 
-  const auto column_views = fct::collect::vector<ColumnViewVariant>(
-      iota | VIEWS::transform(make_column_view));
-
-  const auto string_vectors = fct::collect::vector<std::vector<std::string>>(
-      column_views | VIEWS::transform(to_string_vector));
-
-  const auto data = make_data(string_vectors);
+  const auto data = transpose(
+      fct::collect::vector(column_views | VIEWS::transform(to_string_vector)));
 
   const auto nrows = make_nrows(column_views, _force_nrows);
 
-  Poco::JSON::Object obj;
+  const auto basis = fct::make_field<"draw", std::int32_t>(_draw) *
+                     fct::make_field<"data">(data);
 
-  obj.set("draw", _draw);
-
-  obj.set("data", data);
-
-  if (nrows) {
-    obj.set("recordsTotal", *nrows);
-
-    obj.set("recordsFiltered", *nrows);
+  if (!nrows) {
+    return basis;
   }
 
-  return obj;
+  return containers::DataFrameContent(
+      basis * fct::make_field<"recordsTotal">(*nrows) *
+      fct::make_field<"recordsFiltered">(*nrows));
 }
 
 // ----------------------------------------------------------------------------
 
-containers::DataFrame ViewParser::parse(const Poco::JSON::Object& _obj) {
-  const auto type = JSON::get_value<std::string>(_obj, "type_");
+containers::DataFrame ViewParser::parse(
+    const commands::DataFrameOrView& _cmd) const {
+  const auto handle = [this](const auto& _cmd) -> containers::DataFrame {
+    using Type = std::decay_t<decltype(_cmd)>;
 
-  if (type == "DataFrame") {
-    const auto name = JSON::get_value<std::string>(_obj, "name_");
-    return utils::Getter::get(name, *data_frames_);
-  }
+    if constexpr (std::is_same<Type, DataFrameOp>()) {
+      const auto name = fct::get<"name_">(_cmd);
+      return utils::Getter::get(name, *data_frames_);
+    } else {
+      const auto& base = *fct::get<"base_">(_cmd);
 
-  const auto base = *JSON::get_object(_obj, "base_");
+      auto df = parse(base);
 
-  auto df = parse(base);
+      add_column(_cmd, &df);
 
-  add_column(_obj, &df);
+      drop_columns(_cmd, &df);
 
-  drop_columns(_obj, &df);
+      subselection(_cmd, &df);
 
-  subselection(_obj, &df);
+      df.set_build_history(commands::Fingerprint(_cmd));
 
-  df.set_build_history(Poco::JSON::Object::Ptr(new Poco::JSON::Object(_obj)));
+      return df;
+    }
+  };
 
-  return df;
+  return fct::visit(handle, _cmd.val_);
 }
 
 // ----------------------------------------------------------------------------
 
 std::tuple<containers::DataFrame, std::vector<containers::DataFrame>,
            std::optional<containers::DataFrame>>
-ViewParser::parse_all(const Poco::JSON::Object& _cmd) {
-  const auto to_df =
-      [this](const Poco::JSON::Object::Ptr& _obj) -> containers::DataFrame {
-    assert_true(_obj);
-    return parse(*_obj);
+ViewParser::parse_all(const commands::DataFramesOrViews& _cmd) const {
+  const auto to_df = [this](const auto& _obj) -> containers::DataFrame {
+    return parse(_obj);
   };
 
-  const auto population_obj = JSON::get_object(_cmd, "population_df_");
+  const auto population_obj = _cmd.get<"population_df_">();
 
-  const auto peripheral_objs =
-      JSON::array_to_obj_vector(JSON::get_array(_cmd, "peripheral_dfs_"));
+  const auto peripheral_objs = _cmd.get<"peripheral_dfs_">();
 
-  const auto validation_obj = _cmd.has("validation_df_")
-                                  ? JSON::get_object(_cmd, "validation_df_")
-                                  : Poco::JSON::Object::Ptr();
+  const auto validation_obj = _cmd.get<"validation_df_">();
 
   const auto population = to_df(population_obj);
 
-  const auto peripheral = fct::collect::vector<containers::DataFrame>(
-      peripheral_objs | VIEWS::transform(to_df));
+  const auto peripheral =
+      fct::collect::vector(peripheral_objs | VIEWS::transform(to_df));
 
   const auto validation =
       validation_obj
-          ? std::make_optional<containers::DataFrame>(to_df(validation_obj))
+          ? std::make_optional<containers::DataFrame>(to_df(*validation_obj))
           : std::optional<containers::DataFrame>();
 
   return std::make_tuple(population, peripheral, validation);
@@ -400,60 +385,62 @@ ViewParser::parse_all(const Poco::JSON::Object& _cmd) {
 
 // ----------------------------------------------------------------------------
 
-void ViewParser::subselection(const Poco::JSON::Object& _obj,
+void ViewParser::subselection(const ViewOp& _cmd,
                               containers::DataFrame* _df) const {
-  if (!_obj.has("subselection_")) {
+  if (!_cmd.get<"subselection_">()) {
     return;
   }
 
-  const auto json_col = *JSON::get_object(_obj, "subselection_");
+  const auto handle = [this, &_cmd, _df](const auto& _json_col) {
+    using Type = std::decay_t<decltype(_json_col)>;
 
-  const auto type = JSON::get_value<std::string>(json_col, "type_");
+    if constexpr (std::is_same<Type, commands::BooleanColumnView>()) {
+      const auto column_view =
+          BoolOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col);
 
-  if (type == BOOLEAN_COLUMN_VIEW) {
-    const auto column_view =
-        BoolOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col);
+      const auto data_ptr = column_view.to_vector(0, _df->nrows(), true);
 
-    const auto data_ptr = column_view.to_vector(0, _df->nrows(), true);
+      assert_true(data_ptr);
 
-    assert_true(data_ptr);
+      _df->where(*data_ptr);
+    } else {
+      const auto data_ptr =
+          FloatOpParser(categories_, join_keys_encoding_, data_frames_)
+              .parse(_json_col)
+              .to_vector(0, std::nullopt, false);
 
-    _df->where(*data_ptr);
-  } else {
-    const auto data_ptr =
-        NumOpParser(categories_, join_keys_encoding_, data_frames_)
-            .parse(json_col)
-            .to_vector(0, std::nullopt, false);
+      assert_true(data_ptr);
 
-    assert_true(data_ptr);
+      const auto& key_float = *data_ptr;
 
-    const auto& key_float = *data_ptr;
+      auto key = std::vector<size_t>(key_float.size());
 
-    auto key = std::vector<size_t>(key_float.size());
+      for (size_t i = 0; i < key_float.size(); ++i) {
+        const auto ix_float = key_float[i];
 
-    for (size_t i = 0; i < key_float.size(); ++i) {
-      const auto ix_float = key_float[i];
+        if (ix_float < 0.0) {
+          throw std::runtime_error(
+              "Index on a numerical subselection cannot be "
+              "smaller than zero!");
+        }
 
-      if (ix_float < 0.0) {
-        throw std::runtime_error(
-            "Index on a numerical subselection cannot be "
-            "smaller than zero!");
+        const auto ix = static_cast<size_t>(ix_float);
+
+        if (ix >= _df->nrows()) {
+          throw std::runtime_error(
+              "Index on a numerical subselection out of "
+              "bounds!");
+        }
+
+        key[i] = ix;
       }
 
-      const auto ix = static_cast<size_t>(ix_float);
-
-      if (ix >= _df->nrows()) {
-        throw std::runtime_error(
-            "Index on a numerical subselection out of "
-            "bounds!");
-      }
-
-      key[i] = ix;
+      _df->sort_by_key(key);
     }
+  };
 
-    _df->sort_by_key(key);
-  }
+  return std::visit(handle, *_cmd.get<"subselection_">());
 }
 
 // ----------------------------------------------------------------------------
